Add encryption feature
This does few things, firslty, it does check when a job is created from a scheduledJob, if it's not the case it will update the scheduledJob authentication configuration to encrypt the data once. Lastly, when the job is cloned from the scheduled job it will decrypt the authentication configuration data in that normal job to let the service run the harvest.