Fix and cleanup of attach auth conf query
The query was rather messy and the `WHERE` clause was querying for
information that was not used in the `INSERT` clause. That can be
removed. There was an error in the `WHERE` clause that would cause it to
return nothing, so the `INSERT` could not go through. There was also an
error in the `INSERT` clause where the security configuration would
point to another security configuration instead of the secrets.